Refactored Reviews Endnpoints
