+++ src/controllers/movies.js
@@ -2,12 +2,9 @@ import {movieDbServiceClient} from '../util/api';
 
 /**
  * Movies Controller Initialization Function
- * @param {Object} ControllerParams - Controller Parameters
- * @param {Object} ControllerParams.userModel - Initialized User Model
- * @param {Object} ControllerParams.reviewModel - Initialized Review Model
  * @returns {Object} Controller Object
  */
-export default ({userModel, reviewModel}) => {
+export default () => {
   // Helper Methods
   const sortMovies = movieList =>
     movieList.sort((movieA, movieB) => {
@@ -134,63 +131,6 @@ export default ({userModel, reviewModel}) => {
       return next(error);
     }
   };
-  const reviewMovie = async (req, res, next) => {
-    try {
-      const user = await userModel.findOne({where: {email: req.user.email}});
-
-      if (!user) {
-        const userNotFoundError = new Error(
-          `User with email - ${req.user.email} Not Found`
-        );
-        userNotFoundError.statusCode = 404;
-        throw userNotFoundError;
-      }
-
-      try {
-        await movieDbServiceClient({
-          url: `/movie/${req.body.movieId}`,
-          method: 'get',
-        });
-      } catch (error) {
-        if (error?.response?.data?.status_message.includes('could not be found')) {
-          const movieNotFoundError = new Error(
-            "A movie with that ID can't be found"
-          );
-          movieNotFoundError.statusCode = 404;
-          throw movieNotFoundError;
-        } else {
-          throw error;
-        }
-      }
-
-      let review = await reviewModel.findOne({
-        where: {userId: req.user.id, movieId: req.body.movieId},
-      });
-      let responseStatus = 201;
-
-      if (!review) {
-        review = await reviewModel.create({
-          movieId: req.body.movieId,
-          content: req.body.review,
-          userId: req.user.id,
-        });
-      } else {
-        review = await review.update({
-          content: req.body.review,
-        });
-        responseStatus = 200;
-      }
-
-      return res.status(responseStatus).json({
-        status: 'success',
-        message: 'Movie Reviewed',
-        data: {review},
-      });
-    } catch (error) {
-      if (!error.statusCode) error.statusCode = 500;
-      return next(error);
-    }
-  };
 
   const searchForMovies = async (req, res, next) => {
     try {
@@ -242,7 +182,6 @@ export default ({userModel, reviewModel}) => {
   return {
     getLatestMovies,
     getSingleMovieDetails,
-    reviewMovie,
     searchForMovies,
   };
 };
