+++ src/routes/movies.js
@@ -1,28 +1,16 @@
 import MoviesController from '../controllers/movies';
-import AuthMiddleware from '../middlewares/auth';
 
 /**
  * Movies Router Initialization Function
  * @param  {Object} RouterParams - Router Parameters
  * @param  {Object} RouterParams.express - Express
- * @param {Object} RouterParams.jwt - JsonWebToken
  * @param {Object} RouterParams.expressValidator - Express Validator
  * @param {Object} RouterParams.validator - Custom Validator
- * @param {Object} RouterParams.userModel - Initialized User Model
- * @param {Object} RouterParams.reviewModel - Initialized Review Model
  * @returns {Object} ExpressRouter
  */
-export default ({
-  express,
-  jwt,
-  expressValidator,
-  validator,
-  userModel,
-  reviewModel,
-}) => {
+export default ({express, expressValidator, validator}) => {
   const moviesRouter = express.Router();
-  const authMiddleware = AuthMiddleware({jwt});
-  const moviesController = MoviesController({userModel, reviewModel});
+  const moviesController = MoviesController();
 
   moviesRouter.get(
     '/',
@@ -58,22 +46,6 @@ export default ({
     ],
     validator,
     moviesController.getSingleMovieDetails
-  );
-
-  moviesRouter.post(
-    '/review',
-    authMiddleware.verifyToken,
-    [
-      expressValidator('movieId')
-        .trim()
-        .isNumeric()
-        .withMessage('Movie ID is required'),
-      expressValidator('review')
-        .trim()
-        .isString(),
-    ],
-    validator,
-    moviesController.reviewMovie
   );
 
   return moviesRouter;
